Cleanup moving suggested to separate file
